+++ PID_Py/PID.py
@@ -1,6 +1,7 @@
 import time
 from enum import Flag, auto
 from threading import Thread
+import logging
 
 class HistorianParams(Flag):
     """
@@ -50,10 +51,17 @@ class PID:
             - HistorianParams.PROCESS_VALUE : PID process value
             - HistorianParams.OUTPUT : PID output
     
+    historianLenght: int, default = 100000
+        The maximum lenght of the historian. When the limit is reached, remove the oldest element.
+    
     outputLimits: tuple[float, float], default = (None, None)
         Limit the output between a minimum and a maximum (min, max).
         If a limit is set to None, the limit is deactivated.
         If `outputLimit` is set to None, there is no limits.
+    
+    logger: logging.Logger or str, default = None
+        Logging system. `logging.Logger` instance or logger name (str) can be passed.
+        If it's anything else (None or other type), the PID will not send any log.
     
     Attributes
     ----------
@@ -81,6 +89,9 @@ class PID:
     historian: dict[str, list]
         PID value recorded
     
+    historianLenght: int
+        Same as `historianLenght` in parameters section.
+    
     output: float
         PID output
     
@@ -98,6 +109,9 @@ class PID:
         Bump can occur if the setpoint is too far from process value when automatic mode is reactivated.
         Default value : True
     
+    logger: logging.Logger
+        Contain the `logging.Logger` instance. If it's None or other type, the PID will not send any log.
+    
     Methods
     -------
     compute(processValue, setpoint)
@@ -106,7 +120,7 @@ class PID:
     __call__(processValue, setpoint)
         call `compute`. Is a code simplification.
     """
-    def __init__(self, kp: float, ki: float, kd: float, indirectAction: bool = False, integralLimit: float = None, historianParams: HistorianParams = None, outputLimits: tuple[float, float] = (None, None)) -> None:
+    def __init__(self, kp: float, ki: float, kd: float, indirectAction: bool = False, integralLimit: float = None, historianParams: HistorianParams = None, historianLenght: int = 100000, outputLimits: tuple[float, float] = (None, None), logger: logging.Logger = None) -> None:
         # PID parameters
         self.kp = kp
         self.ki = ki
@@ -126,6 +140,11 @@ class PID:
         # Historian setup
         self.historianParams = historianParams
         self.historian = {}
+
+        if (historianLenght <= 0):
+            raise ValueError("`historianLenght` can't be 0 or negative!")
+
+        self.historianLenght = historianLenght
         
         if HistorianParams.P in self.historianParams:
             self.historian["P"] = []
@@ -162,7 +181,21 @@ class PID:
 
         # Output
         self.output = 0.0
-    
+
+        # Logger
+        self.logger = None
+        if (isinstance(logger, logging.Logger)):
+            self.logger = logger
+            self.logger.info("PID object created")
+        elif (isinstance(logger, str)):
+            self.logger = logging.getLogger(logger)
+            self.logger.info("PID object created")
+
+        self.memManualMode = False
+        self.memIntegralLimit = False
+        self.outputLimitsReached = False
+        self.memoutputLimitsReached = False
+
     def compute(self, processValue: float, setpoint: float) -> float:
         """
         PID calculation execution
@@ -180,6 +213,14 @@ class PID:
         float
             Return the PID output (same as `self.output`)
         """
+        # Logging mode switching
+        if (self.manualMode and not self.memManualMode):
+            self.logger.info("PID switched to manual mode")
+        elif (not self.manualMode and self.memManualMode):
+            self.logger.info("PID switched to automatic mode")
+        
+        self.memManualMode = self.manualMode
+        
         if self._startTime is not None and self._lastTime is not None:
             actualTime = time.time()
 
@@ -210,8 +251,25 @@ class PID:
                 if self.integralLimit is not None:
                     if self._i > self.integralLimit:
                         self._i = self.integralLimit
+
+                        # Integral limit warning
+                        if (not self.memIntegralLimit):
+                            self.logger.warning("Integral part has reached the limit (%d)", self.integralLimit)
+                            self.memIntegralLimit = True
+                        
                     elif self._i < -self.integralLimit:
                         self._i = -self.integralLimit
+
+                        # Integral limit warning
+                        if (not self.memIntegralLimit):
+                            self.logger.warning("Integral part has reached the limit (-%d)", self.integralLimit)
+                            self.memIntegralLimit = True
+                    
+                    # Reset integral limit warning memory
+                    else:
+                        self.memIntegralLimit = False
+                else:
+                    self.memIntegralLimit = False
                 
                 # ===== Derivative part =====
                 self._d = ((error - self._lastError) / deltaTime) * self.kd
@@ -225,39 +283,75 @@ class PID:
 
                 # ========== End of automatic mode ==========
 
+            self.outputLimitsReached = False
+
             # Output limitation
             if self.outputLimits is not None:
                 if self.outputLimits[0] is not None:
                     if _output < self.outputLimits[0]:
                         _output = self.outputLimits[0]
+
+                        self.outputLimitsReached = True
                 if self.outputLimits[1] is not None:
                     if _output > self.outputLimits[1]:
                         _output = self.outputLimits[1]
+                        
+                        self.outputLimitsReached = True
+            
+            # Output limit reached warning message
+            if (self.outputLimitsReached and not self.memoutputLimitsReached):
+                self.logger.warning("Output limits reached (%d, %d)", self.outputLimits[0], self.outputLimits[1])
+            
+            self.memoutputLimitsReached = self.outputLimitsReached
 
             # ===== Historian =====
             if HistorianParams.P in self.historianParams:
                 self.historian["P"].append(self._p)
+                
+                if len(self.historian["P"]) > self.historianLenght:
+                    del self.historianLenght["P"][0]
             
             if HistorianParams.I in self.historianParams:
                 self.historian["I"].append(self._i)
+                
+                if len(self.historian["I"]) > self.historianLenght:
+                    del self.historianLenght["I"][0]
 
             if HistorianParams.D in self.historianParams:
                 self.historian["D"].append(self._d)
+                
+                if len(self.historian["D"]) > self.historianLenght:
+                    del self.historianLenght["D"][0]
             
             if HistorianParams.OUTPUT in self.historianParams:
                 self.historian["OUTPUT"].append(_output)
+                
+                if len(self.historian["OUTPUT"]) > self.historianLenght:
+                    del self.historianLenght["OUTPUT"][0]
             
             if HistorianParams.SETPOINT in self.historianParams:
                 self.historian["SETPOINT"].append(setpoint)
+                
+                if len(self.historian["SETPOINT"]) > self.historianLenght:
+                    del self.historianLenght["SETPOINT"][0]
             
             if HistorianParams.PROCESS_VALUE in self.historianParams:
                 self.historian["PROCESS_VALUE"].append(processValue)
+                
+                if len(self.historian["PROCESS_VALUE"]) > self.historianLenght:
+                    del self.historianLenght["PROCESS_VALUE"][0]
 
             if HistorianParams.ERROR in self.historianParams:
                 self.historian["ERROR"].append(error)
+                
+                if len(self.historian["ERROR"]) > self.historianLenght:
+                    del self.historianLenght["ERROR"][0]
 
             if (HistorianParams.P in self.historianParams) or (HistorianParams.I in self.historianParams) or (HistorianParams.D in self.historianParams) or (HistorianParams.ERROR in self.historianParams) or (HistorianParams.OUTPUT in self.historianParams) or (HistorianParams.PROCESS_VALUE in self.historianParams) or (HistorianParams.SETPOINT in self.historianParams):
                 self.historian["TIME"].append(actualTime - self._startTime)
+                
+                if len(self.historian["TIME"]) > self.historianLenght:
+                    del self.historianLenght["TIME"][0]
             
             # ===== Saving data for next execution =====
             self._lastError = error
@@ -325,11 +419,18 @@ class ThreadedPID(PID, Thread):
             - HistorianParams.SETPOINT : PID setpoint
             - HistorianParams.PROCESS_VALUE : PID process value
             - HistorianParams.OUTPUT : PID output
+            
+    historianLenght: int, default = 100000
+        The maximum lenght of the historian. When the limit is reached, remove the oldest element.
     
     outputLimits: tuple[float, float], default = (None, None)
         Limit the output between a minimum and a maximum (min, max).
         If a limit is set to None, the limit is deactivated.
         If `outputLimit` is set to None, there is no limits.
+    
+    logger: logging.Logger or str, default = None
+        Logging system. `logging.Logger` instance or logger name (str) can be passed.
+        If it's anything else (None or other type), the PID will not send any log.
     
     cycleTime: float, default = 0.0
         Define the minimum time between two PID calculations.
@@ -355,8 +456,8 @@ class ThreadedPID(PID, Thread):
     start()
         Used to start the thread.
     """
-    def __init__(self, kp: float, ki: float, kd: float, indirectAction: bool = False, integralLimit: float = None, historianParams: HistorianParams = None, outputLimits: tuple[float, float] = (None, None), cycleTime: float = 0.0) -> None:
-        PID.__init__(self, kp, ki, kd, indirectAction, integralLimit, historianParams, outputLimits)
+    def __init__(self, kp: float, ki: float, kd: float, indirectAction: bool = False, integralLimit: float = None, historianParams: HistorianParams = None, historianLenght: int = 100000, outputLimits: tuple[float, float] = (None, None), logger: logging.Logger = None, cycleTime: float = 0.0) -> None:
+        PID.__init__(self, kp, ki, kd, indirectAction, integralLimit, historianParams, historianLenght, outputLimits, logger)
         Thread.__init__(self)
 
         self.setpoint = 0.0