+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change log
 
+## [0.2.2] - 2023-01-27
+### Added
+- Logging system added
+- Historian maximum lenght added
+
+### Modififed
+
+### Fixed
+
 ## [0.2.1] - 2023-01-26
 ### Added
 - Threaded PID class documentation